StealthCamera.pde
This is the process android app, which is written by me. This one simply takes the picture in stealth mode and saves it to the phone memory. 

This is a simple code, you can run the sketch in Processing IDE and load it your mobile phone. 

Things not included in the code :

- No audio/mic input is included
- No Video